Creation of Home Section
Creation of Home Section assembling two components and one sub-component: BackgroundVid, which displays a background video (video_03_gaussian_blur) that fills the entire screen and behind of that renders a background image (image_1) in case the video fails to load; AnimatedRole, a sub-component which contains logic to simulate a typing animation in a string passed as props; and Fullname, which renders a main title with a Fade animation and displays the AnimatedRole sub-component underneath. The three components are imported in an index.js file and then exported to be used in Home.jsx. The App.js file has been modified to render the new Home.jsx component