Changing style rules and improving code
Adding color, fontFamilies, fontSizes and sizes rules to styles.js and applying them in ProfilePic, NavLinks, Footer, SideBar.jsx, BackgroundVid, Fullname and Home.jsx components; changing items props in SideBar to be an array of objects instead of an array of arrays, changing the items PropTypes in NavLinks to define the received object via PropTypes.shape, separating the states in AnimatedRole component into two: state and blink. State contains and object with related states that works together, and blink an individual state which determines when the cursor needs to start blinking. The state of 'typingSpeed' was removed and replaced by a variable called 'speed'. Adding new webm video format to BackgroundVid.