Creation of ContactMe Section
Creation of ContactMe Section assembling three global components and two local components: InputTextField, which styles a Material UI's textfield; SectionTitle, which displays the title of the section with an icon; SectionSubtitle, which displays a given subtitle when needed; ContactInfo, which renders contact information in three aligned blocks, with an icon; and AssembledForm, which imports InputTextField and renders three instances of it, with a styled submit button. More changes were made to other components, like: modifying App.js to render ContactMe.jsx, adding logic to IconMaker to render new icons and variate its sizes, importing the new icons on IconMaker/index.js to be used on IconMaker, adding style rules to style.js, adding icons to ICONS object in constants.js, creating index.js in ContactMe/components to import all the components used in ContactMe.jsx and installing Formik dependencies.