Creation of AboutMe Section
Creation of AboutMe Section assembling three global components and two local components: the globals are SectionTitle, SectionSubtitle, and StyledButton, a component that renders a blue button; the locals are ListMaker, which takes several values and renders a list, and ServicesInfo, which renders an icon with a title and text, all related to a specific service. An index.js was created to import all these components and be exported on ContactMe.jsx. Another changes were made, like: importing more icons on IconMaker/index.js and exporting them in IconMaker, adding logic on IconMaker to render the new icons and adding keywords in constants.js related to the icons.