Creation of Resume Section
Creation of Resume Section Assembling two global and two local components. The globals are SectionTitle and SectionSubtitle; the locals are TimelineMaker, which displays a timeline separated in lapses, where each one has a title, a subtitle, a paragraph and a date; and SkillMeter, which renders a given number of progress bars with an animation. All this components where imported in an index.js file to be exported and used in Resume.jsx